+++ src/main/java/com/newz/api/news/keywordnews/service/KeywordNewsService.java
@@ -1,0 +1,10 @@
+package com.newz.api.news.keywordnews.service;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class KeywordNewsService {
+
+}
